Explicitly implment is_unicast_link_local
This avoids the otherwise confusing errors in the shape of

> Error binding socket to [fe80::...]:0: Invalid argument (os error 22)

Code for this was already in place but pending on is_unicast_link_local
stabilization. As this has not happened in several years, a manual
workaround (with the same note to wait for stabilization) is chosen.

Workaround-For: https://github.com/rust-lang/rust/issues/27709